Added helper method, called method to show card
I was able to play any card based on the rank and suit of the top card; Helper method should help with fixing that logical error. When game starts, the top card isn't shown, nor is the next top card when the player plays a card ; called .reveal() when printing those cards to show in terminal.